Went through code with Youssef and added additional comments
Removed default samples

Recorded clips now are added to the current beat when the recording
starts, not when it stops.

Checks to see if audio exists for a beat and should be played is no
longer done by checking if PlayArray contains a -1, but if SamplesArray
contains a null value instead of a Sample object.